ETHBE-688
Syncer have to fail each time when app initialization is failed.
It is impossible because gunicorn will restart application after each failing.

There is opinion what gunicorn is not nessecary thing in docker container.
We remove it and add clean cli to run we API.

>>> jsearch --log-level INFO --no-json-formatter api -p 8000

Example higher will start API on port 8000 without json formatter and
with log level `INFO`